chore: make webpack config file separate
