+++ package.json
@@ -112,13 +112,14 @@
     "webpack": "^4.40.2",
     "webpack-bundle-analyzer": "^3.5.0",
     "webpack-cli": "^3.3.9",
-    "webpack-dev-server": "^3.8.1"
+    "webpack-dev-server": "^3.8.1",
+    "webpack-merge": "^4.2.2"
   },
   "scripts": {
     "test": "karma start --no-single-run",
     "test:ne": "KARMA_SERVER=ne karma start",
     "test:types": "tsc --project test/types",
-    "serve": "webpack-dev-server --mode development",
+    "serve": "webpack-dev-server --config webpack.dev.config.js",
     "cpy-dist2doc": "mkdir -p doc/dist && cp -f -r dist doc",
     "release-note": "tuie",
     "eslint": "eslint ./src/ts ./test-ts --ext ts --ext tsx",
@@ -126,10 +127,10 @@
     "doc:dev": "tuidoc --dev",
     "doc": "tuidoc",
     "build": "rm -rf dist && npm run build:es5 && npm run build:es6 && npm run build:esm && node tsBannerGenerator.js",
-    "build:es5": "webpack --mode production && webpack --mode production --minify",
-    "build:es6": "webpack --mode production --es6 && webpack --mode production --minify --es6",
+    "build:es5": "webpack && webpack --minify",
+    "build:es6": "webpack --config webpack.es6.config.js && webpack --config webpack.es6.config.js --minify",
     "build:esm": "ttsc -p tsconfig.esm.json && prettier dist/esm/**/*.js --write",
-    "analyze": "webpack --mode production --minify --profile --json > stats.json && webpack-bundle-analyzer stats.json ./dist",
+    "analyze": "webpack --minify --profile --json > stats.json && webpack-bundle-analyzer stats.json ./dist",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
